Moved lock code to its own directory
I also added a smart lock manager class for handling sets of locks. I
started to realize that we may have to rethink how we handle sets of
locks. If we use the push button notification option, we run the risk of
getting like 5 of those if you have 5 locks in the house.