Major code refactoring
Unable to test this code, so it is very possible that it doesn't work. I
converted the lock manager list of locks over to a hash map to avoid
n-squared comparisons every time we add a lock. However, a lot of code
had to be reworked to make this happen. I'm expecting saving and loading
to fail.