Forced only adafruit devices to show up on device list
User can no longer find devices that aren't adafruit devices. This would
cause a bug where the app would treat it as a lock and create a symbol
for it. It would be better to possibly wait for communication from the
lock to confirm it was a lock before storing it, but this was a quick
fix under impending deadlines.